Move click-actions drill machinery from visualizations into querying
The Mode class, getMode/getEmbeddingMode, the mode definitions
(Default/Archived/List), and the querying-dependent actions
(CombineColumnsAction, ExtractColumnAction, NativeQueryClickFallback)
are the drill layer: chart events in, query transformations out. That
bridge lives above the lower endpoint, so it moves from visualizations
(P0) into querying (P1). Actions with no querying dependency
(ColumnFormattingAction, HideColumnAction) stay in visualizations.

Consumers codemodded, no shims. Visualization.tsx temporarily imports
the machinery from querying; that edge is removed by the getMode lift
later in this stack.
