+++ frontend/src/metabase/querying/click-actions/Mode/Mode.ts
@@ -1,15 +1,15 @@
 import { transformClickedDataPoint } from "metabase/embedding-sdk/lib/transform-clicked";
 import type { MetabasePluginsConfig } from "metabase/embedding-sdk/types/plugins";
 import { queryDrill } from "metabase/querying/drills/utils/query-drill";
-import type { DrillThruDisplayInfo } from "metabase-lib";
-import type Question from "metabase-lib/v1/Question";
-
 import type {
   ClickAction,
   ClickActionsMode,
   ClickObject,
   QueryClickActionsMode,
-} from "../../types";
+} from "metabase/visualizations/types";
+import type { DrillThruDisplayInfo } from "metabase-lib";
+import type Question from "metabase-lib/v1/Question";
+
 
 export class Mode implements ClickActionsMode {
   _question: Question;
