+++ frontend/src/metabase/querying/click-actions/Mode/index.ts
