+++ frontend/src/metabase/querying/click-actions/actions/CombineColumnsAction/index.ts
