+++ frontend/src/metabase/querying/click-actions/actions/NativeQueryClickFallback.tsx
