+++ frontend/src/metabase/querying/click-actions/modes/ArchivedMode.ts
@@ -1,4 +1,4 @@
-import type { QueryClickActionsMode } from "../../types";
+import type { QueryClickActionsMode } from "metabase/visualizations/types";
 
 export const ArchivedMode: QueryClickActionsMode = {
   name: "archived",
